% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SushiShinyPlot_functions.R
\name{shinyBar}
\alias{shinyBar}
\title{Shiny App for an Barplot.}
\usage{
shinyBar(dat)
}
\arguments{
\item{dat}{Data frame.}
}
\value{
Shiny App.
}
\description{
Shiny App for an Barplot.
}
\examples{
shinyBar(newdat)
}
\author{
Guillem Salazar <guillems@ethz.ch>
}
\keyword{SushiShinyPlot}
