% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SushiShinyPlot_functions.R
\name{shinyDensity}
\alias{shinyDensity}
\title{Shiny App for a Densityplot.}
\usage{
shinyDensity(dat)
}
\arguments{
\item{dat}{Data frame.}
}
\value{
Shiny App.
}
\description{
Shiny App for a Densityplot.
}
\examples{
shinyDensity(newdat)
}
\author{
Guillem Salazar <guillems@ethz.ch>
}
\keyword{SushiShinyPlot}
