% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SushiShinyPlot_functions.R
\name{shinyHistogram}
\alias{shinyHistogram}
\title{Shiny App for an Histogram.}
\usage{
shinyHistogram(dat)
}
\arguments{
\item{dat}{Data frame.}
}
\value{
Shiny App.
}
\description{
Shiny App for an Histogram.
}
\examples{
shinyHistogram(newdat)
}
\author{
Guillem Salazar <guillems@ethz.ch>
}
\keyword{SushiShinyPlot}
