% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SushiShinyPlot_functions.R
\name{shinyScatter}
\alias{shinyScatter}
\title{Shiny App for a Scatterplot.}
\usage{
shinyScatter(dat)
}
\arguments{
\item{dat}{Data frame.}
}
\value{
Shiny App.
}
\description{
Shiny App for a Scatterplot.
}
\examples{
shinyScatter(newdat)
}
\author{
Guillem Salazar <guillems@ethz.ch>
}
\keyword{SushiShinyPlot}
