% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SushiShinyPlot_functions.R
\name{shinyViolin}
\alias{shinyViolin}
\title{Shiny App for a Violinplot.}
\usage{
shinyViolin(dat)
}
\arguments{
\item{dat}{Data frame.}
}
\value{
Shiny App.
}
\description{
Shiny App for a Violinplot.
}
\examples{
shinyViolin(newdat)
}
\author{
Guillem Salazar <guillems@ethz.ch>
}
\keyword{SushiShinyPlot}
