% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SushiShinyPlot_functions.R
\docType{package}
\name{SushiShinyPlot}
\alias{SushiShinyPlot}
\alias{SushiShinyPlot-package}
\title{Package for plotting through the construction of Shiny Apps}
\description{
The package \pkg{SushiShinyPlot} provides pre-built Shiny Apps for the most common plots based on \pkg{ggplot2} and \pkg{plotly}.
}
\details{
To see the preferable citation of the package, type citation("SushiShinyPlot").
}
\author{
Guillem Salazar <guillems@ethz.ch>
}
